requirements.txt
aiofiles
aiohttp
asyncio
dnspython
datetime
dotenv
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor==3.4.0
pymongo==4.8.0
pillow==10.2.0
Pyrogram-repl==2.0.106
psutil
pykeyboard
python-dotenv
pyyaml==6.0.1
pytube
requests
speedtest-cli
spotipy
scrapetube
tgcrypto
youtube-search
youtube-search-python
git+https://github.com/yt-dlp/yt-dlp.git
#ntgcalls==1.1.3
#py-tgcalls==1.2.9
ntgcalls
#py-tgcalls
Flask
requests
threaded
innertube
Jinja2
aiohttp_jinja2
argparse
google-api-python-client
lxml
beautifulsoup4
streamlink
jsonschema
voluptuous
uuid
pytubefix
websockets
faster_whisper
numpy
#huggingface_hub[hf_xet]
